fix: dual webkit and correctness fix
Add fragmentID argument to getResponseJSON().
Since DoActionUtils.processActivityResult() can signal the webkit, 
we should clear the dispatch and action strings prior to the signal
just in case the signal gets processed on the webkit UI thread
before we return from this method (unclear if the UI threads are the same).