fix: double-load of ODK Survey webkit
(1) Remove the onResume() of ODKSurveyWebView. The reload of the page is handled by databaseAvailable().  This call was causing a double-load of the page.
(2) Add onPause() and delegate it to the webkit within the fragment. This ensures that the should-reload flag is set so that when a doAction response comes back, the webpage will reload. 
(3) Remove getWebKitResourceId() from application -- with the databaseAvailable() pathway and onPause() delegation, this (hack) is no longer needed.
(4) don't adjust the hash of a page via gotoUrlHash() if it has been loaded. This was another redundant call that should be removed.
(5) update JS zip with changes.